1) is there a bug? std::memcpy(tdRequest.data, GetData(), Constants::MAX_DATA_SIZE); 2) implement correct way to stop server and client threads; 3) auto_ptr is deprecated. please replace it. also it would be good to replace DWORD type with something from C++11; 4) please, add some routines which will test your task, e.g. via MS Test.

+++ client-server/request.h
@@ -11,9 +11,9 @@ typedef struct tagTDATA
     // Priority of request 0-255 (0 - highest priority)
     BYTE cPriority;
     // Request's creation time in system clock ticks
-    DWORD dwTicks;
+    unsigned long dwTicks;
     // Unique client ID
-    DWORD dwClientId;
+    unsigned long dwClientId;
     // Abstract data
     char data[Constants::MAX_DATA_SIZE];
 } TDATA, *PTDATA;
