+++ client-server/view.h
@@ -8,29 +8,29 @@
 class View
 {
 public:
-	View(DWORD clients);
+	View(unsigned long clients);
 
 	~View()
 	{
 		SetConsoleCursorPosition(m_console, COORD{ 0, m_bottomY });
 	}
 
-	void ShowRequest(const TDATA &request, DWORD size);
-	void ShowLog(const TDATA &request, DWORD size);
+	void ShowRequest(const TDATA &request, unsigned long size);
+	void ShowLog(const TDATA &request, unsigned long size);
 
 private:
 	HANDLE m_console;
-	DWORD m_clients;
+	unsigned long m_clients;
 	SHORT m_topY;
 	SHORT m_topX;
 	SHORT m_bottomY;
 
-	void ShowSize(DWORD size);
+	void ShowSize(unsigned long size);
 	void Paint(const std::vector<std::wstring> &figure, const COORD &coord);
 	void PaintServer();
 	void PaintLog();
 	void PaintQueue();
-	void PaintClient(DWORD id, const COORD &coord);
+	void PaintClient(unsigned long id, const COORD &coord);
 	void PaintClients();
 	void PaintConnects();
 	void Write(const std::wstring &msg, const COORD &coord);
@@ -40,7 +40,7 @@ private:
 	COORD GetLogFlushCoord() const;
 	COORD GetQueueCoord() const;
 	COORD GetSizeFlushCoord() const;
-	COORD GetClientCoord(DWORD id) const;
-	COORD GetConnectsCoord(DWORD id) const;
+	COORD GetClientCoord(unsigned long id) const;
+	COORD GetConnectsCoord(unsigned long id) const;
 
 };
